requirements.txt
openai
pyttsx3
speechrecognition
python-dotenv
pyaudio
### Mac OS X (os.platform == 'Darwin')
pyobjc>=2.4
#pyobjc
gTTS
pyautogui
pydub






